fix generate_randombytes

+++ DiscreetCore/ripemd160/ripemd160.h
@@ -23,9 +23,9 @@ typedef struct {
 } ripemd160_ctx;
 
 EXPORT void ripemd160_init(ripemd160_ctx * md);
-EXPORT void ripemd160_update(ripemd160_ctx *self, const unsigned char *p, unsigned long length);
+EXPORT void ripemd160_update(ripemd160_ctx *self, const unsigned char *p, unsigned long long length);
 EXPORT void ripemd160_final(ripemd160_ctx *self, unsigned char *out);
-EXPORT void ripemd160(const void *in, unsigned long inlen, void *out);
+EXPORT void ripemd160(const void *in, unsigned long long inlen, void *out);
 
 #ifdef __cplusplus
 }
